creados tests de MatrixController y de GlobalExceptionHandler. Configurado para que la aplicación no compile si no pasa el 80% de cobertura de tests

+++ src/main/java/com/salvador/droneControl/infrastructure/controller/MatrixController.java
@@ -40,7 +40,7 @@ public class MatrixController {
     @GetMapping("/get/{id}")
     public ResponseEntity<Matrix> getMatrixById(@Parameter(description = "ID de la matriz a buscar", example = "1") @PathVariable long id) {
         logger.info("Obteniendo matriz de id: {}", id);
-        Matrix matrix = matrixService.getMatrixById((int) id);
+        Matrix matrix = matrixService.getMatrixById(id);
         return new ResponseEntity<>(matrix, HttpStatus.OK);
     }
